requirements.txt
Flask==2.3.3 
requests==2.31.0 
beautifulsoup4==4.12.3 
google-generativeai==0.7.2 
ratelimit==2.2.1 
gunicorn==22.0.0 
Werkzeug==3.0.1